tests: snap tests as a component
