+++ tests/snap_testing/__init__.py
@@ -1,0 +1,3 @@
+"""
+FPGAd snap tests for CLI interactions
+"""
