+++ tests/snap_testing/common/__init__.py
@@ -1,0 +1,1 @@
+"""Common utilities for FPGA snap tests."""
