+++ tests/snap_testing/test_default/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for default behavior (without platform specification)."""
