+++ tests/snap_testing/test_universal/__init__.py
@@ -1,0 +1,1 @@
+"""Universal platform tests."""
