+++ tests/snap_testing/test_xlnx/__init__.py
@@ -1,0 +1,1 @@
+"""Xilinx platform tests."""
